New folder/bookstore-api/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0
werkzeug==2.3.7